[FEAT] AddPage implementation
